+++ src/main.js
@@ -10,6 +10,7 @@ export { default as Input } from './components/Input';
 export { default as Label } from './components/Label';
 export { default as OverlayWrapper } from './components/OverlayWrapper';
 export { default as Popover } from './components/Popover';
+export { default as Poptip } from './components/Poptip';
 export { default as Radio } from './components/Radio';
 export { default as Table } from './components/Table';
 export { default as TabNav } from './components/TabNav';