+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 
 ## [Unreleased][Unreleased]
 ### Added
+- [Feature] Add a `behavior` prop to `OverlayWrapper` that determines if the overlay appears on click or hover.
+- [Feature] Add a `Poptip` component that uses `OverlayWrapper`. (#519)
 - [Feature] Add a `shouldHideOnClick` prop to `OverlayWrapper` to prevent it from always closing when clicking outside the overlay. (#654)
 - [Feature] Add `onShow` and `onHide` props to `OverlayWrapper`. (#653)
 - [Feature] Add a `centerGroup` prop to `ButtonRow`. (#651)